+++ crates/gmsol/src/cli/mod.rs
@@ -1,0 +1,4 @@
+/// Wallet utils.
+pub mod wallet;
+
+pub use self::wallet::signer_from_source;