+++ crates/gmsol/tests/integration_test/mod.rs
@@ -1,1 +1,7 @@
+/// Setup.
+pub mod setup;
 
+/// Test Client.
+pub mod test_client;
+
+mod oracle;